Added GPUBuffer and Geometry classes
Need to implement the draw function for the geometry class.
Will need to create a global bindings class/object/thing in order to fetch current bindings for various bound objects.
Switched to a level down before continuing with shader properties. Felt like I was creating a tool in a void with no direction.