Bug fixing, we have something being presented
Everything seems to be wrong.
Presentation is not waiting on a wait semaphore, so I added a fence. It works.
The render pass images seem to be changing layout for some reason after presentation, even though they're not the images being presented.
My usage of CmdBlitImage is not consistent with the specification and yet the validation layers seem to not be complaining.